internal/tray: refactor tray handling into its own package
It's not as clean a separation as I'd like, but it'll do for now.
