+++ internal/tray/status-icon-active.png
