+++ internal/tray/status-icon-inactive.png
