+++ internal/ui/app.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"deedles.dev/mk"
+	"deedles.dev/trayscale/internal/tray"
 	"deedles.dev/trayscale/internal/tsutil"
 	"deedles.dev/trayscale/internal/version"
 	"deedles.dev/trayscale/internal/xcmp"
@@ -42,7 +43,7 @@ type App struct {
 	app      *adw.Application
 	win      *MainWindow
 	settings *gio.Settings
-	tray     *tray
+	tray     *tray.Tray
 
 	statusPage *adw.StatusPage
 	peerPages  map[key.NodePublic]*peerPage
@@ -277,16 +278,16 @@ func (a *App) initSettings(ctx context.Context) {
 		switch key {
 		case "tray-icon":
 			if a.settings.Boolean("tray-icon") {
-				go startSystray(func() { a.initTray(ctx) })
+				go tray.Start(func() { a.initTray(ctx) })
 				return
 			}
-			stopSystray()
+			tray.Stop()
 		}
 	})
 
 init:
 	if (a.settings == nil) || a.settings.Boolean("tray-icon") {
-		go startSystray(func() { a.initTray(ctx) })
+		go tray.Start(func() { a.initTray(ctx) })
 	}
 }
 
@@ -394,7 +395,7 @@ func (a *App) onAppActivate(ctx context.Context) {
 
 func (a *App) initTray(ctx context.Context) {
 	if a.tray == nil {
-		a.tray = initTray(a.online)
+		a.tray = tray.New(a.online)
 	}
 
 	for {
@@ -415,7 +416,7 @@ func (a *App) initTray(ctx context.Context) {
 
 // Quit exits the app completely, causing Run to return.
 func (a *App) Quit() {
-	stopSystray()
+	tray.Stop()
 	a.app.Quit()
 }
